Move feed definition to own file and add rel link to base template
