+++ wagtailsite/models.py
@@ -1,8 +1,7 @@
-from datetime import date, datetime, time
+from datetime import date
 
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.core.management import call_command
-from django.contrib.syndication.views import Feed
 from django.db import models
 from django.dispatch import receiver
 from django.http import HttpResponse
@@ -198,27 +197,3 @@ BlogPage.promote_panels = [
     ImageChooserPanel('feed_image'),
     FieldPanel('tags'),
 ]
-
-
-class BlogFeed(Feed):
-    title = "The Wagtail Blog"
-    link = "/blog/"
-    description = "The latest news and views from the Wagtail team"
-
-    def items(self):
-        return BlogPage.objects.live().order_by('-date')
-
-    def item_title(self, item):
-        return item.title
-
-    def item_description(self, item):
-        return item.intro if item.intro else item.body
-
-    def item_link(self, item):
-        return item.full_url
-
-    def item_author_name(self, item):
-        pass
-
-    def item_pubdate(self, item):
-        return datetime.combine(item.date, time())
